+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added tabular output for long form outputs
+
+- Added sorting option for file size
+
+- Added sorting option for creation time
+
+- Added option to reverse sort output
+
+- Added option to display numeric ids instead of named ones
+
+### Fixed
+
+- Fixed panic when path does not exist.
 
 ## 0.0.1 - Initial release